Reordered EXTI reset code in HAL_GPIO_DeInit()
Scenario:
Firmware configures a GPIO for EXTI line interrupt on falling edge. In
hardware the GPIO is pulled to Vdd by a pull-up resistor. After the
falling edge is detected (single instance), the firmware reconfigures
the GPIO pin for a different purpose.

Problem:
EXTI line interrupt is triggered twice, while in fact only one falling
edge is generated on the I/O pin.

Root cause:
Improper EXTI reset order in HAL_GPIO_DeInit(). After the first time the
EXTI line interrupt is fired, the firmware calls HAL_GPIO_DeInit() to
de-initialize the GPIO pin because the resource is not needed anymore.
Because the way the code is ordered in HAL_GPIO_DeInit(), '1'-to-'0'
transition might be arise in the "Input data register" because the
mode is reset to Analog High Impedance mode. This disable the internal
Schmitt trigger, which continuously produces a '0' on its output.
If the I/O was pulled high by pull-up resistor, a '1'-to-'0' transition
triggers the EXTI line edge detector and fires the IRQ because the EXTI
line configuration registers are not yet reset at this point in
HAL_GPIO_DeInit().

Solution:
Reorder the EXTI reset code in HAL_GPIO_DeInit() to disable the EXIT
line interrupt before switching the value of EXTI line multiplexer and
resetting the mode of the GPIO pin to analog high-impedance mode.

Fixes #1.